Documentation

+++ ez.h
@@ -21,14 +21,16 @@ Size generateSize(uint16_t x,uint16_t y){
 
 RGBA generateRGBA(uint8_t r, uint8_t g, uint8_t b, uint8_t a);
 double clamp(double d, double min, double max);
-float mapRGBA(double x);
+float mapRGBA(uint8_t x);
 void windowColor(uint8_t r,uint8_t b,uint8_t g,uint8_t a);
 
 int setupGLAD();
 int setupWindow(Size size, char* name);
 void setupViewport(Size size);
 
-
+/**
+ * Runs glfwInit() and some WindowHints (You may need to change them by yourself)
+ */
 void setupGLFW(){
     glfwInit();
     glfwWindowHint(GLFW_CONTEXT_VERSION_MINOR, 3);
@@ -41,6 +43,11 @@ void setEzWindow(GLFWwindow *x){
     window = x;
 }
 
+/**
+ * Initializing the window
+ * @param size The Size type that has x and y for scaling
+ * @param name Window title
+ */
 int setupWindow(Size size, char* name){
     GLFWwindow *window = glfwCreateWindow(size.x, size.y, name, NULL, NULL);
     if (window == NULL)
@@ -53,6 +60,10 @@ int setupWindow(Size size, char* name){
     setEzWindow(window);
 }
 
+
+/**
+ * Setups glad and returns -1 if it fails
+ */
 int setupGLAD(){
     if (!gladLoadGLLoader((GLADloadproc)glfwGetProcAddress))
     {
@@ -66,19 +77,37 @@ void framebuffer_size_callback(GLFWwindow *window, int width, int height)
     glViewport(0, 0, width, height);
 }
 
+/**
+ * Setups the viewport of the glViewport
+ * @param size The Size type that includes the x and y for scaling
+ * Also calls the framebuffer_size_callback for feedback (from ez.h)
+ */
 void setupViewport(Size size){
     glViewport(0, 0, size.x, size.y);
     glfwSetFramebufferSizeCallback(window, framebuffer_size_callback);
 }
 
+
+/**
+ * Generating RGBA typedef
+ * @param all unsigned int, a number between 0 and 255
+ */
 RGBA generateRGBA(uint8_t r, uint8_t g, uint8_t b, uint8_t a) {
     return (RGBA){r,g,b,a};
 }
 
+/**
+ * Convert RGBA type to floatRGBA type
+ * @param color The RGBA type that will get converted
+ */
 floatRGBA generateFloatRGBA(RGBA color){
     return (floatRGBA){mapRGBA(color.r),mapRGBA(color.g),mapRGBA(color.b),mapRGBA(color.a)};
 }
 
+/**
+ * Set window background with simple RGBA
+ * @param all RGBA parameters which require a number between 0 and 255
+ */
 void windowColorRGBA(uint8_t r,uint8_t g,uint8_t b,uint8_t a)
 {
     floatRGBA color1 = generateFloatRGBA(generateRGBA(r,g,b,a));
@@ -87,10 +116,24 @@ void windowColorRGBA(uint8_t r,uint8_t g,uint8_t b,uint8_t a)
     glfwSwapBuffers(window);
 }
 
+/**
+ * Simple math function to map a number
+ * @param x The number you want to map
+ * @param inMin Minimum value of your number
+ * @param inMax Maxmimum value of your number
+ * @param outMin Minimum value of output
+ * @param outMax Maximum value of output
+ */
 double map(double x, double inMin, double inMax, double outMin, double outMax) {
     return outMin + (x - inMin) * (outMax - outMin) / (inMax - inMin);
 }
 
-float mapRGBA(double x){
+/**
+ * Map an RGBA parameter (uint8_t), which is between 0 and 255 to be between 0 and 1
+ * 
+ * @param x uint8_t thats needed to be converted
+ * 
+ */
+float mapRGBA(uint8_t x){
     return (float)map(x,0,255,0,1);
 }
